RatEnemy_Beginnings
Added rat enemy, the animation is terrible rn so I have to fix that, also there were some changes to the spear enemy. I also tweaked the code in the flock script (I added two lines lol) which makes it so we can have as many boids at the start of the game as we want and the value is also editable in the inspector.